GCC 11 build (#6)
* attempt gcc11

* Fix timing in python

+++ main.py
@@ -5,7 +5,7 @@ def chess_bot(obs):
     f="./Alexandria" if os.path.exists("./Alexandria") else "/kaggle_simulations/agent/Alexandria"
     if not p or p.poll()!=None:
         p=subprocess.Popen(f,stdin=subprocess.PIPE,stdout=subprocess.PIPE,stderr=subprocess.PIPE,text=1)
-    t=obs.remainingOverageTime*1000
+    t=int(obs.remainingOverageTime*1000)
     # Treat delay as inc
     p.stdin.write(f"position fen {obs.board}\ngo wtime {t} btime {t} winc 70 binc 70\n")
     p.stdin.flush()
